Move users collection to root namespace
