+++ server/portfolio/collections.ts
@@ -1,4 +1,3 @@
 export const PORTFOLIO_SKILLS = 'portfolio_skills';
 export const PORTFOLIO_BADGES = 'portfolio_badges';
 export const PORTFOLIO_PROJECTS = 'portfolio_projects';
-export const PORTFOLIO_USERS = 'portfolio_users';
