+++ server/portfolio/schemas.ts
@@ -90,21 +90,3 @@ export const portfolioProjectsSchema = {
     },
   },
 };
-
-export const portfolioUsersSchema = {
-  $jsonSchema: {
-    bsonType: 'object',
-    required: ['username', 'password', 'role'],
-    properties: {
-      username: {
-        bsonType: 'string',
-      },
-      password: {
-        bsonType: 'string',
-      },
-      role: {
-        bsonType: 'string',
-      },
-    },
-  },
-};
