+++ server/root/collections.ts
@@ -1,0 +1,1 @@
+export const USERS = 'users';
