+++ server/root/users.ts
@@ -1,7 +1,7 @@
 import bcrypt from 'bcryptjs';
 import { Router } from 'express';
 import { database } from '../database';
-import { PORTFOLIO_USERS } from './collections';
+import { USERS } from './collections';
 import { generateJwt } from '../auth';
 import { BadRequestError, UnauthorizedError } from '../utils';
 import { User } from '../types';
@@ -10,7 +10,7 @@ const usersRouter = Router();
 
 usersRouter.route('/users/login').post(async (req, res, next) => {
   try {
-    const collection = database.collection<User>(PORTFOLIO_USERS);
+    const collection = database.collection<User>(USERS);
     const { username, password } = req.body;
     const user = await collection.findOne({ username });
     const isValidUser = user && bcrypt.compareSync(password, user.password);
@@ -26,7 +26,7 @@ usersRouter.route('/users/login').post(async (req, res, next) => {
 
 usersRouter.route('/users/register-admin').post(async (req, res, next) => {
   try {
-    const collection = database.collection<User>(PORTFOLIO_USERS);
+    const collection = database.collection<User>(USERS);
     const existingDocuments = await collection.countDocuments();
     if (existingDocuments > 0) {
       throw new BadRequestError('Admin user already exists');
